VCDA-496: Fix org case insensitive issue
Some commands like vcd vdc list/use were failing because we don't do case insensitive org name comparison.
This can happen when we create a Org with uppercase, say ORG1,
login in vcd-cli using lower case( org1) and then try to list the vdc in ORG1.
Since the org_name from the api does not match with the in_use_org we store in the profiles file, we get in the issue.