[VCDA-603] Fixed 'vcd catalog info' when user can't access ACL settings of the catalog. (#225)
Non admin users who are not owner of a catalog but has viewing rights, can't access the ACL of the catalog. The command vcd catalog info disregards this limitation and tries to retreive the ACL of catalog for all users. This leads to an exception being raised, and such users can never use the command.

As a fix we omit catalog ACL details for users who can't access them.