Add ui extensions commands
Following commads are added: generate, list, deploy, delete.
They will help the user to manage his ui extensions.

CLI Spinner class is implemented to indicate the action of the user.

Extension generator is class which inherits
generator abstract class and
give the ability to copy existing template from
the file system, of the user, and populate it
with the data, given by the user.

Prompt and PromptLauncher classes are
build on top of python click lib to give
the ability to validate and collect
the data from the user.
Signed-off-by: Nikola Vladimirov Iliev <nvladimirovi@vmware.com>