Under-score changed to Hyphen in windows resulting in gateway_tests failure in unix
Before the fix, output for command vcd gateway -h shows configure_external_network in unix but configure-external-network in windows. Because of this, test case with is
posting command with '-' was succeeding in windows but failing in unix.

Same is with sub-allocate-ip sub-group.

After this fix, both group mentioned above will work with '-' and also show in 'vcd gateway -h' output.

Testing Done: Run the command where '-' is shown as the output and also executed the command without any error.
vcd gateway configure-external network -h