VP-2395 : Fix - click version issue of VCD-CLI.
This CLN contains changes according to
vcd-cli 21.1.0 have issue.
If this will run on click version of 6.7 or less then it will not work.
It will work perfectly with click version 7.0 because
code changes in https://jira.eng.vmware.com/browse/VP-2153 contain hidden attribute of click that was added in 7.0.