Add --network-quota to vdc create
Adds --network-quota integer option to the
  vcd vdc create
command, which sets the network_quota setting in the
underlying call to org.create_org_vdc - this is the Maximum
number of network objects that can be deployed in this vdc.

The default value is zero (like the underlying call), which
(unike all the similar settings) means that *no* network
objects can be created within the vdc.

Signed-off-by: Nigel Metheringham <nigel.metheringham@redcentricplc.com>