vcd-cli: login: Add login with API Access Token
Add support for authentication using API Access Tokens as described in
[Generate an API Access Token][1]. The authentication follows RFC6749
(OAuth 2.0). If the `user` argument of the login cli command is set to
"api_token" the `session-id` is interpreted as API Access Token. An
oAuth 2.0 authentication request is made to the token endpoint and the
access_token from the response is used as `session-id`. Furthermore, the
pyvcloud client's `rehydrate_from_token` function must be called with
the optional argument `is_jwt_token=True`.

API Access Token authentication is useful when vCloud Director is
configured to authenticate through an [external identity provider][2].

[1]: https://docs.vmware.com/en/VMware-Cloud-Director/10.4/VMware-Cloud-Director-Tenant-Portal-Guide/GUID-A1B3B2FA-7B2C-4EE1-9D1B-188BE703EEDE.html
[2]: https://docs.vmware.com/en/VMware-Cloud-Director/10.5/VMware-Cloud-Director-Service-Provider-Admin-Guide/GUID-3326986B-931C-4FDE-AF47-D5A863191072.html

Signed-off-by: Martin Fleischer <mfleischer@evoila.de>